Make IEntity an abstract class, move all Entity members to it.
