+++ Robust.Shared/GameObjects/IEntity.cs
@@ -5,200 +5,231 @@ using Robust.Shared.Network;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization.Manager.Attributes;
 using Robust.Shared.Timing;
+using Robust.Shared.Utility;
+using Robust.Shared.ViewVariables;
 
 namespace Robust.Shared.GameObjects
 {
     [CopyByRef]
-    public interface IEntity
+    public abstract class IEntity
     {
-        GameTick LastModifiedTick { get; internal set; }
+        #region Members
 
-        /// <summary>
-        /// The Entity Manager that controls this entity.
-        /// </summary>
-        IEntityManager EntityManager { get; }
+        /// <inheritdoc />
+        public IEntityManager EntityManager { get; }
 
-        /// <summary>
-        ///     The name of this entity.
-        ///     This is the actual IC display name.
-        /// </summary>
-        string Name { get; set; }
+        /// <inheritdoc />
+        [ViewVariables]
+        public EntityUid Uid { get; }
 
-        /// <summary>
-        ///     The unique ID of this entity.
-        ///     Unique IDs are unique per entity,
-        ///     and correspond to counterparts across the network.
-        /// </summary>
-        EntityUid Uid { get; }
+        public EntityLifeStage LifeStage { get => MetaData.EntityLifeStage; internal set => MetaData.EntityLifeStage = value; }
 
-        /// <summary>
-        ///     The current lifetime stage of this entity. You can use this to check
-        ///     if the entity is initialized or being deleted.
-        /// </summary>
-        EntityLifeStage LifeStage { get; internal set; }
+        [ViewVariables]
+        public GameTick LastModifiedTick { get => MetaData.EntityLastModifiedTick; internal set => MetaData.EntityLastModifiedTick = value; }
 
-        /// <summary>
-        ///     Whether this entity has fully initialized.
-        /// </summary>
-        bool Initialized { get; }
 
-        bool Initializing { get; }
+        /// <inheritdoc />
+        [ViewVariables]
+        public EntityPrototype? Prototype
+        {
+            get => MetaData.EntityPrototype;
+            internal set => MetaData.EntityPrototype = value;
+        }
 
-        /// <summary>
-        ///     True if the entity has been deleted.
-        /// </summary>
-        bool Deleted { get; }
+        /// <inheritdoc />
+        [ViewVariables(VVAccess.ReadWrite)]
+        public string Description
+        {
+            get => MetaData.EntityDescription;
+            set => MetaData.EntityDescription = value;
+        }
 
-        bool Paused { get; set; }
+        /// <inheritdoc />
+        [ViewVariables(VVAccess.ReadWrite)]
+        public string Name
+        {
+            get => MetaData.EntityName;
+            set => MetaData.EntityName = value;
+        }
 
-        /// <summary>
-        ///     The prototype that was used to create this entity.
-        /// </summary>
-        EntityPrototype? Prototype { get; }
+        /// <inheritdoc />
+        public bool Initialized => LifeStage >= EntityLifeStage.Initialized;
 
-        /// <summary>
-        /// The string that describes this entity via examine
-        /// </summary>
-        string Description { get; set; }
+        /// <inheritdoc />
+        public bool Initializing => LifeStage == EntityLifeStage.Initializing;
 
-        /// <summary>
-        ///     Determines if this entity is still valid.
-        /// </summary>
-        /// <returns>True if this entity is still valid.</returns>
-        bool IsValid();
+        /// <inheritdoc />
+        public bool Deleted => LifeStage >= EntityLifeStage.Deleted;
 
-        /// <summary>
-        ///     The entity's transform component.
-        /// </summary>
-        TransformComponent Transform { get; }
+        [ViewVariables]
+        public bool Paused { get => MetaData.EntityPaused; set => MetaData.EntityPaused = value; }
 
-        /// <summary>
-        ///     The MetaData Component of this entity.
-        /// </summary>
-        MetaDataComponent MetaData { get; }
+        private TransformComponent? _transform;
+
+        /// <inheritdoc />
+        [ViewVariables]
+        public TransformComponent Transform => _transform ??= GetComponent<TransformComponent>();
+
+        private MetaDataComponent? _metaData;
+
+        /// <inheritdoc />
+        [ViewVariables]
+        public MetaDataComponent MetaData
+        {
+            get => _metaData ??= GetComponent<MetaDataComponent>();
+            internal set => _metaData = value;
+        }
+
+        #endregion Members
+
+        #region Initialization
+
+        public IEntity(IEntityManager entityManager, EntityUid uid)
+        {
+            EntityManager = entityManager;
+            Uid = uid;
+        }
+
+        /// <inheritdoc />
+        public bool IsValid()
+        {
+            return EntityManager.EntityExists(Uid);
+        }
+
+        #endregion Initialization
+
+        #region Component Messaging
+
+        /// <inheritdoc />
+        [Obsolete("Component Messages are deprecated, use Entity Events instead.")]
+        public void SendMessage(IComponent? owner, ComponentMessage message)
+        {
+            var components = EntityManager.GetComponents(Uid);
+            foreach (var component in components)
+            {
+                if (owner != component)
+                    component.HandleMessage(message, owner);
+            }
+        }
+
+        /// <inheritdoc />
+        [Obsolete("Component Messages are deprecated, use Entity Events instead.")]
+        public void SendNetworkMessage(IComponent owner, ComponentMessage message, INetChannel? channel = null)
+        {
+            EntityManager.EntityNetManager?.SendComponentNetworkMessage(channel, this, owner, message);
+        }
+
+        #endregion Component Messaging
+
+        #region Components
 
         /// <summary>
         ///     Public method to add a component to an entity.
         ///     Calls the component's onAdd method, which also adds it to the component manager.
         /// </summary>
-        /// <typeparam name="T">The component type to add.</typeparam>
-        /// <returns>The newly added component.</returns>
-        T AddComponent<T>()
-            where T : Component, new();
+        /// <param name="component">The component to add.</param>
+        public void AddComponent(Component component)
+        {
+            EntityManager.AddComponent(this, component);
+        }
 
-        /// <summary>
-        ///     Removes the component with the specified reference type,
-        ///     Without needing to have the component itself.
-        /// </summary>
-        /// <typeparam name="T">The component reference type to remove.</typeparam>
-        void RemoveComponent<T>();
+        /// <inheritdoc />
+        public T AddComponent<T>()
+            where T : Component, new()
+        {
+            return EntityManager.AddComponent<T>(this);
+        }
 
-        /// <summary>
-        ///     Checks to see if the entity has a component of the specified type.
-        /// </summary>
-        /// <typeparam name="T">The component reference type to check.</typeparam>
-        /// <returns>True if the entity has a component of type <typeparamref name="T" />, false otherwise.</returns>
-        bool HasComponent<T>();
+        /// <inheritdoc />
+        public void RemoveComponent<T>()
+        {
+            EntityManager.RemoveComponent<T>(Uid);
+        }
 
-        /// <summary>
-        ///     Checks to see ift he entity has a component of the specified type.
-        /// </summary>
-        /// <param name="type">The component reference type to check.</param>
-        /// <returns></returns>
-        bool HasComponent(Type type);
+        /// <inheritdoc />
+        public bool HasComponent<T>()
+        {
+            return EntityManager.HasComponent<T>(Uid);
+        }
 
-        /// <summary>
-        ///     Retrieves the component of the specified type.
-        /// </summary>
-        /// <typeparam name="T">The component reference type to fetch.</typeparam>
-        /// <returns>The retrieved component.</returns>
-        /// <exception cref="Shared.GameObjects.UnknownComponentException">
-        ///     Thrown if there is no component with the specified type.
-        /// </exception>
-        T GetComponent<T>();
+        /// <inheritdoc />
+        public bool HasComponent(Type type)
+        {
+            return EntityManager.HasComponent(Uid, type);
+        }
 
-        /// <summary>
-        ///     Retrieves the component of the specified type.
-        /// </summary>
-        /// <param name="type">The component reference type to fetch.</param>
-        /// <returns>The retrieved component.</returns>
-        /// <exception cref="Shared.GameObjects.UnknownComponentException">
-        ///     Thrown if there is no component with the specified type.
-        /// </exception>
-        IComponent GetComponent(Type type);
+        /// <inheritdoc />
+        public T GetComponent<T>()
+        {
+            DebugTools.Assert(!Deleted, "Tried to get component on a deleted entity.");
 
-        /// <summary>
-        ///     Attempt to retrieve the component with specified type,
-        ///     writing it to the <paramref name="component" /> out parameter if it was found.
-        /// </summary>
-        /// <typeparam name="T">The component reference type to attempt to fetch.</typeparam>
-        /// <param name="component">The component, if it was found. Null otherwise.</param>
-        /// <returns>True if a component with specified type was found.</returns>
-        bool TryGetComponent<T>([NotNullWhen(true)] out T? component) where T : class;
+            return (T)EntityManager.GetComponent(Uid, typeof(T));
+        }
 
-        /// <summary>
-        ///     Attempt to retrieve the component with specified type,
-        ///     returning it if it was found.
-        /// </summary>
-        /// <typeparam name="T">The component reference type to attempt to fetch.</typeparam>
-        /// <returns>The component, if it was found. Null otherwise.</returns>
-        T? GetComponentOrNull<T>() where T : class;
+        /// <inheritdoc />
+        public IComponent GetComponent(Type type)
+        {
+            DebugTools.Assert(!Deleted, "Tried to get component on a deleted entity.");
 
-        /// <summary>
-        ///     Attempt to retrieve the component with specified type,
-        ///     writing it to the <paramref name="component" /> out parameter if it was found.
-        /// </summary>
-        /// <param name="type">The component reference type to attempt to fetch.</param>
-        /// <param name="component">The component, if it was found. Null otherwise.</param>
-        /// <returns>True if a component with specified type was found.</returns>
-        bool TryGetComponent(Type type, [NotNullWhen(true)] out IComponent? component);
+            return EntityManager.GetComponent(Uid, type);
+        }
 
-        /// <summary>
-        ///     Attempt to retrieve the component with specified type,
-        ///     returning it if it was found.
-        /// </summary>
-        /// <param name="type">The component reference type to attempt to fetch.</param>
-        /// <returns>The component, if it was found. Null otherwise.</returns>
-        IComponent? GetComponentOrNull(Type type);
+        /// <inheritdoc />
+        public bool TryGetComponent<T>([NotNullWhen(true)] out T? component) where T : class
+        {
+            DebugTools.Assert(!Deleted, "Tried to get component on a deleted entity.");
 
-        /// <summary>
-        ///     Queues this entity for deletion at the end of the tick.
-        /// </summary>
-        void QueueDelete();
+            return EntityManager.TryGetComponent(Uid, out component);
+        }
 
-        /// <summary>
-        ///     Deletes this entity.
-        /// </summary>
-        void Delete();
+        public T? GetComponentOrNull<T>() where T : class
+        {
+            return TryGetComponent(out T? component) ? component : default;
+        }
 
-        /// <summary>
-        ///     Returns all components on the entity.
-        /// </summary>
-        /// <returns>An enumerable of components on the entity.</returns>
-        IEnumerable<IComponent> GetAllComponents();
+        /// <inheritdoc />
+        public bool TryGetComponent(Type type, [NotNullWhen(true)] out IComponent? component)
+        {
+            DebugTools.Assert(!Deleted, "Tried to get component on a deleted entity.");
 
-        /// <summary>
-        ///     Returns all components that are assignable to <typeparamref name="T"/>.
-        ///     This does not go by component references.
-        /// </summary>
-        /// <typeparam name="T">The type that components must implement.</typeparam>
-        /// <returns>An enumerable over the found components.</returns>
-        IEnumerable<T> GetAllComponents<T>();
+            return EntityManager.TryGetComponent(Uid, type, out component);
+        }
 
-        /// <summary>
-        ///     Sends a message to all other components in this entity.
-        /// </summary>
-        /// <param name="owner">Object that sent the event.</param>
-        /// <param name="message">Message to send.</param>
-        [Obsolete("Component Messages are deprecated, use Entity Events instead.")]
-        void SendMessage(IComponent? owner, ComponentMessage message);
+        public IComponent? GetComponentOrNull(Type type)
+        {
+            return TryGetComponent(type, out var component) ? component : null;
+        }
 
-        /// <summary>
-        ///     Sends a message over the network to the counterpart component. This works both ways.
-        /// </summary>
-        /// <param name="owner"></param>
-        /// <param name="message">Message to send.</param>
-        [Obsolete("Component Messages are deprecated, use Entity Events instead.")]
-        void SendNetworkMessage(IComponent owner, ComponentMessage message, INetChannel? channel = null);
+        /// <inheritdoc />
+        public void QueueDelete()
+        {
+            EntityManager.QueueDeleteEntity(this);
+        }
+
+        /// <inheritdoc />
+        public void Delete()
+        {
+            EntityManager.DeleteEntity(this);
+        }
+
+        /// <inheritdoc />
+        public IEnumerable<IComponent> GetAllComponents()
+        {
+            return EntityManager.GetComponents(Uid);
+        }
+
+        /// <inheritdoc />
+        public IEnumerable<T> GetAllComponents<T>()
+        {
+            return EntityManager.GetComponents<T>(Uid);
+        }
+
+        #endregion Components
+
+        /// <inheritdoc />
+        public override string ToString()
+        {
+            return EntityManager.ToPrettyString(Uid);
+        }
     }
 }
